+++ instrumentation/bunny/test/test_helper.rb
@@ -19,4 +19,5 @@ OpenTelemetry::SDK.configure do |c|
   c.error_handler = ->(exception:, message:) { raise(exception || message) }
   c.logger = Logger.new($stderr, level: ENV.fetch('OTEL_LOG_LEVEL', 'fatal').to_sym)
   c.add_span_processor SPAN_PROCESSOR
+  c.propagators = [OpenTelemetry::Trace::Propagation::TraceContext.text_map_propagator]
 end