+++ instrumentation/dalli/lib/opentelemetry/instrumentation/dalli/instrumentation.rb
@@ -20,7 +20,7 @@ module OpenTelemetry
         end
 
         option :peer_service, default: nil, validate: :string
-        option :db_statement, default: :include, validate: %I[omit obfuscate include]
+        option :db_statement, default: :obfuscate, validate: %I[omit obfuscate include]
 
         private