+++ instrumentation/delayed_job/lib/opentelemetry/instrumentation/delayed_job/plugins/tracer_plugin.rb
@@ -14,13 +14,13 @@ module OpenTelemetry
         class TracerPlugin < Delayed::Plugin
           class << self
             def instrument_enqueue(job, &block)
-              return block.call(job) unless enabled?
+              return yield(job) unless enabled?
 
               attributes = build_attributes(job)
-              attributes['messaging.operation'] = 'send'
+              attributes['messaging.operation'] = 'publish'
               attributes.compact!
 
-              tracer.in_span("#{job_queue(job)} send", attributes: attributes, kind: :producer) do |span|
+              tracer.in_span("#{job_queue(job)} publish", attributes: attributes, kind: :producer) do |span|
                 yield job
                 span.set_attribute('messaging.message_id', job.id.to_s)
                 add_events(span, job)
@@ -28,7 +28,7 @@ module OpenTelemetry
             end
 
             def instrument_invoke(job, &block)
-              return block.call(job) unless enabled?
+              return yield(job) unless enabled?
 
               attributes = build_attributes(job)
               attributes['messaging.delayed_job.attempts'] = job.attempts if job.attempts
@@ -56,7 +56,6 @@ module OpenTelemetry
             end
 
             def add_events(span, job)
-              span.add_event('created_at', timestamp: job.created_at)
               span.add_event('run_at', timestamp: job.run_at) if job.run_at
               span.add_event('locked_at', timestamp: job.locked_at) if job.locked_at
             end
@@ -84,8 +83,10 @@ module OpenTelemetry
           end
 
           callbacks do |lifecycle|
+            # rubocop:disable Performance/MethodObjectAsBlock
             lifecycle.around(:enqueue, &method(:instrument_enqueue))
             lifecycle.around(:invoke_job, &method(:instrument_invoke))
+            # rubocop:enable Performance/MethodObjectAsBlock
           end
         end
       end