+++ instrumentation/faraday/lib/opentelemetry/instrumentation/faraday/middlewares/tracer_middleware.rb
@@ -44,9 +44,9 @@ module OpenTelemetry
           def span_creation_attributes(http_method:, url:)
             instrumentation_attrs = {
               'http.method' => http_method,
-              'http.url' => OpenTelemetry::Common::Utilities.cleanse_url(url.to_s),
-              'net.peer.name' => url.host
+              'http.url' => OpenTelemetry::Common::Utilities.cleanse_url(url.to_s)
             }
+            instrumentation_attrs['net.peer.name'] = url.host if url.host
             config = Faraday::Instrumentation.instance.config
             instrumentation_attrs['peer.service'] = config[:peer_service] if config[:peer_service]
             instrumentation_attrs.merge!(
@@ -60,7 +60,7 @@ module OpenTelemetry
 
           def trace_response(span, response)
             span.set_attribute('http.status_code', response.status)
-            span.status = OpenTelemetry::Trace::Status.error unless (100..399).include?(response.status.to_i)
+            span.status = OpenTelemetry::Trace::Status.error unless (100..399).cover?(response.status.to_i)
           end
         end
       end