+++ instrumentation/gruf/example/controllers/example_api_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative '../base'
+
+class ExampleApiController < Gruf::Controllers::Base
+  bind Proto::Example::ExampleAPI::Service
+
+  def example
+    Proto::Example::ExampleResponse.new(response_name: 'Done')
+  end
+end