+++ instrumentation/active_record/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release History: opentelemetry-instrumentation-active_record
 
+### v0.7.0 / 2023-11-22
+
+* BREAKING CHANGE: Drop Rails 6.0 EOL
+
+* ADDED: Drop Rails 6.0 EOL
+
+### v0.6.3 / 2023-10-16
+
+* FIXED: Add Rails 7.1 compatibility
+
+### v0.6.2 / 2023-08-14
+
+* FIXED: Ensure that transaction name property is used, rather than self
+
 ### v0.6.1 / 2023-06-05
 
 * FIXED: Base config options