+++ instrumentation/net_http/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History: opentelemetry-instrumentation-net_http
 
+### v0.22.4 / 2023-11-23
+
+* CHANGED: Applied Rubocop Performance Recommendations [#727](https://github.com/open-telemetry/opentelemetry-ruby-contrib/pull/727)
+
+### v0.22.3 / 2023-11-22
+
+* FIXED: Update `Net::HTTP` instrumentation to no-op on untraced contexts
+
 ### v0.22.2 / 2023-07-21
 
 * ADDED: Update `opentelemetry-common` from [0.19.3 to 0.20.0](https://github.com/open-telemetry/opentelemetry-ruby-contrib/pull/537)