+++ instrumentation/sidekiq/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History: opentelemetry-instrumentation-sidekiq
 
+### v0.25.0 / 2023-09-07
+
+* FIXED: Align messaging instrumentation operation names
+
+### v0.24.4 / 2023-08-07
+
+* FIXED: Allow traces inside jobs while avoiding Redis noise
+
+### v0.24.3 / 2023-08-03
+
+* FIXED: Remove inline linter rules
+
 ### v0.24.2 / 2023-07-21
 
 * ADDED: Update `opentelemetry-common` from [0.19.3 to 0.20.0](https://github.com/open-telemetry/opentelemetry-ruby-contrib/pull/537)