+++ propagator/xray/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History: opentelemetry-propagator-xray
 
+### v0.22.1 / 2023-11-23
+
+* CHANGED: Applied Rubocop Performance Recommendations [#727](https://github.com/open-telemetry/opentelemetry-ruby-contrib/pull/727)
+
 ### v0.22.0 / 2023-04-17
 
 * BREAKING CHANGE: Drop support for EoL Ruby 2.7